rename JSONResponse module to "response", re #2611
